Move credentials in config
