+++ config/auth_pat.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/databricks/databricks-sdk-go/credentials"
+	"github.com/databricks/databricks-sdk-go/config/credentials"
 )
 
 type PatCredentials struct {
@@ -15,7 +15,7 @@ func (c PatCredentials) Name() string {
 	return "pat"
 }
 
-func (c PatCredentials) Configure(ctx context.Context, cfg *Config) (credentials.CredentialsProvider, error) {
+func (c PatCredentials) Configure(ctx context.Context, cfg *Config) (CredentialsProvider, error) {
 	if cfg.Token == "" || cfg.Host == "" {
 		return nil, nil
 	}
